Release 0.4.1 with clean spec directory
Prior to this commit, the spec directory was populated in the 0.4.0
release artifact with a fixtures directory containing unresolvable
symbolic links. This is created during the normal spec testing
process but should be cleaned (`rake spec_clean`) prior to release.

This commit merely increments the module version to 0.4.1 to
represent a new release with a clean spec folder. No behavior has
changed except that.